+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/Internal/Shared/PrometheusExporter.cs
@@ -25,7 +25,10 @@ internal sealed class PrometheusExporter : BaseExporter<Metric>, IPullMetricExpo
 
         this.ScopeInfoEnabled = options.ScopeInfoEnabled;
         this.ScrapeResponseCacheDurationMilliseconds = options.ScrapeResponseCacheDurationMilliseconds;
+        this.TargetInfoEnabled = options.TargetInfoEnabled;
         this.DisableTotalNameSuffixForCounters = options.DisableTotalNameSuffixForCounters;
+        this.ResourceConstantLabels = options.ResourceConstantLabels;
+        this.MaxScrapeResponseSizeBytes = options.MaxScrapeResponseSizeBytes;
 
         this.CollectionManager = new PrometheusCollectionManager(this);
     }
@@ -45,9 +48,15 @@ internal sealed class PrometheusExporter : BaseExporter<Metric>, IPullMetricExpo
 
     internal bool ScopeInfoEnabled { get; }
 
+    internal bool TargetInfoEnabled { get; }
+
     internal int ScrapeResponseCacheDurationMilliseconds { get; }
 
     internal bool DisableTotalNameSuffixForCounters { get; }
+
+    internal Func<string, bool>? ResourceConstantLabels { get; }
+
+    internal int MaxScrapeResponseSizeBytes { get; }
 
     internal Resource Resource
     {